+++ app/models/concerns/domain/discardable.rb
@@ -1,16 +1,6 @@
 module Concerns::Domain::Discardable
   extend ActiveSupport::Concern
 
-  def discard
-    raise 'Domain is already discarded' if discarded?
-
-    statuses << DomainStatus::DELETE_CANDIDATE
-    transaction do
-      save(validate: false)
-      delete_later
-    end
-  end
-
   def keep
     statuses.delete(DomainStatus::DELETE_CANDIDATE)
     transaction do
@@ -21,5 +11,15 @@ module Concerns::Domain::Discardable
 
   def discarded?
     statuses.include?(DomainStatus::DELETE_CANDIDATE)
+  end
+
+  private
+
+  def discard
+    statuses << DomainStatus::DELETE_CANDIDATE
+    transaction do
+      save(validate: false)
+      delete_later
+    end
   end
 end
