+++ test/integration/epp/domain/domain_update_test.rb
@@ -32,9 +32,7 @@ class EppDomainUpdateTest < ApplicationIntegrationTest
   end
 
   def test_discarded_domain_cannot_be_updated
-    travel_to Time.zone.parse('2010-07-05 10:30')
-    @domain.delete_at = Time.zone.parse('2010-07-05 10:00')
-    @domain.discard
+    @domain.update!(statuses: [DomainStatus::DELETE_CANDIDATE])
 
     request_xml = <<-XML
       <?xml version="1.0" encoding="UTF-8" standalone="no"?>
